Fix chatGLM_int4 Error (#586)
* Fix chatGLM_int4 Error

Error: expected scalar type Half but found Float.

* Update models.py

---------

+++ modules/models.py
@@ -234,7 +234,7 @@ class ChatGLM_Client(BaseLLMModel):
         if quantified:
             model = AutoModel.from_pretrained(
                 model_source, trust_remote_code=True
-            ).float()
+            ).half()
         else:
             model = AutoModel.from_pretrained(
                 model_source, trust_remote_code=True
